// examples/soak-testing/src/types.ts
export interface RuntimeCall {
    accounts?:            CallMessage;
    uniqueness?:          null;
    bank?:                CallMessage2;
    sequencer_registry?:  CallMessage3;
    attester_incentives?: CallMessage4Class | CallMessage4Enum;
    prover_incentives?:   CallMessage5Class | CallMessage5Enum;
    example_module?:      CallMessage6;
    chain_state?:         null;
    blob_storage?:        null;
    paymaster?:           CallMessage7;
}

/**
 * Represents the available call messages for interacting with the sov-accounts module.
 *
 * Inserts a new credential id for the corresponding Account.
 */
export interface CallMessage {
    insert_credential_id: string;
}

/**
 * Register an attester, the parameter is the bond amount
 *
 * Register a challenger, the parameter is the bond amount
 *
 * Increases the balance of the attester.
 */
export interface CallMessage4Class {
    register_attester?:   number;
    register_challenger?: number;
    deposit_attester?:    number;
}

/**
 * Start the first phase of the two-phase exit process
 *
 * Finish the two phase exit
 *
 * Exit a challenger
 */
export enum CallMessage4Enum {
    BeginExitAttester = "begin_exit_attester",
    ExitAttester = "exit_attester",
    ExitChallenger = "exit_challenger",
}

/**
 * This enumeration represents the available call messages for interacting with the sov-bank
 * module.
 *
 * Creates a new token with the specified name and initial balance.
 *
 * Transfers a specified amount of tokens to the specified address.
 *
 * Burns a specified amount of tokens.
 *
 * Mints a specified amount of tokens.
 *
 * Freezes a token so that the supply is frozen
 */
export interface CallMessage2 {
    create_token?: CreateToken;
    transfer?:     Transfer;
    burn?:         Burn;
    mint?:         Mint;
    freeze?:       Freeze;
}

export interface Burn {
    /**
     * The amount of tokens to burn.
     */
    coins: Coins;
    [property: string]: any;
}
// ...
export interface CreateToken {
    /**
     * Admins list.
     */
    admins: string[];
    /**
     * The initial balance of the new token.
     */
    initial_balance: number;
    /**
     * The address of the account that the new tokens are minted to.
     */
    mint_to_address: string;
    /**
     * The supply cap of the new token, if any.
     */
    supply_cap?: number | null;
    /**
     * The number of decimal places this token's amounts will have.
     */
    token_decimals?: number | null;
    /**
     * The name of the new token.
     */
    token_name: string;
    [property: string]: any;
}

export interface Freeze {
    /**
     * Address of the token to be frozen
     */
    token_id: string;
    [property: string]: any;
}

export interface Mint {
    /**
     * The amount of tokens to mint.
     */
    coins: Coins;
    /**
     * Address to mint tokens to
     */
    mint_to_address: string;
    [property: string]: any;
}

export interface Transfer {
    /**
     * The amount of tokens to transfer.
     */
    coins: Coins;
    /**
     * The address to which the tokens will be transferred.
     */
    to: string;
    [property: string]: any;
}

/**
 * This enumeration represents the available call messages for interacting with the
 * `ExampleModule` module. The `derive` for [`schemars::JsonSchema`] is a requirement of
 * [`sov_modules_api::ModuleCallJsonSchema`].
 */
export interface CallMessage6 {
    set_value: number;
}

export interface UpdatePolicy {
    payer:  string;
    update: CallMessage7;
    [property: string]: any;
}
// ...
export interface CallMessage7 {
    register_paymaster?:      RegisterPaymaster;
    set_payer_for_sequencer?: SetPayerForSequencer;
    update_policy?:           UpdatePolicy;
}

export interface RegisterPaymaster {
    policy: PaymasterPolicyInitializer;
    [property: string]: any;
}
// ...
export interface SetPayerForSequencer {
    payer: string;
    [property: string]: any;
}

/**
 * Add a new prover as a bonded prover.
 *
 * Increases the balance of the prover, transferring the funds from the prover account to
 * the rollup.
 */
export interface CallMessage5Class {
    register?: number;
    deposit?:  number;
}

/**
 * Unbonds the prover.
 */
export enum CallMessage5Enum {
    Exit = "exit",
}
// ...
export interface CallMessage3 {
    register?:            Register;
    deposit?:             Deposit;
    initiate_withdrawal?: InitiateWithdrawal;
    withdraw?:            Withdraw;
}

export interface Deposit {
    /**
     * The amount to increase.
     */
    amount: number;
    /**
     * The DA address of the sequencer.
     */
    da_address: string;
    [property: string]: any;
}

export interface InitiateWithdrawal {
    /**
     * The DA address of the sequencer you're removing.
     */
    da_address: string;
    [property: string]: any;
}

export interface Register {
    /**
     * The initial balance of the sequencer.
     */
    amount: number;
    /**
     * The Da address of the sequencer you're registering.
     */
    da_address: string;
    [property: string]: any;
}

export interface Withdraw {
    /**
     * The DA address of the sequencer you're removing.
     */
    da_address: string;
    [property: string]: any;
}